Timer widget done and working
