+++ apps/entry/forms.py
@@ -1,38 +1,8 @@
 from django import forms
 from django.core.exceptions import ValidationError
-from django.core.validators import RegexValidator
-from django.forms import widgets, MultiWidget
-from django.template import loader
-from django.utils.safestring import mark_safe
+from django.forms import widgets
 from apps.entry.models import *
-
-class TickerWidget(forms.Widget):
-    template_name = 'entry/components/widgets/ticker.html'
-    image = ''
-
-    def __init__(self, image=''):
-        super().__init__()
-        self.image = image
-
-    def get_context(self, name, value, attrs=None, **kwargs):
-        return {'widget': {
-            'name': name,
-            'value': value,
-            'image': self.image
-        }}
-
-    def render(self, name, value, attrs=None, **kwargs):
-        context = self.get_context(name, value, attrs, **kwargs)
-        template = loader.get_template(self.template_name).render(context)
-        return mark_safe(template)
-
-
-class BooleanWidget(TickerWidget):
-    def format_value(self, value):
-        value = value[0]
-        return super(self, value)
-
-    template_name = 'entry/components/widgets/boolean.html'
+from apps.entry.widgets import *
 
 
 class MatchScoutForm(forms.Form):
@@ -105,6 +75,7 @@ class MatchScoutForm(forms.Form):
     # DEFENSE
     defense_played = forms.BooleanField(widget=BooleanWidget(), label='Played Defense?', required=False)
     team_defended = forms.IntegerField(widget=widgets.NumberInput, required=False)  # TODO Can this now become a popup?
+    defense_time = forms.IntegerField(widget=StopWatchWidget(), label="Defense Stopwatch", initial=0)
     defense_rating = forms.IntegerField(widget=widgets.Select(choices=[
                     (1, "foo"),
                     (2, "bar")
